backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
anyio==4.9.0
asgiref==3.8.1
async-timeout==5.0.1
asyncio==3.4.3
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.7
django-settings-file==2.10.1
exceptiongroup==1.2.2
frozenlist==1.5.0
graypy==2.1.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
multidict==6.1.0
packaging==24.2
persian==0.5.0
pip-review==1.3.0
propcache==0.3.0
python-telegram-bot==22.0
sniffio==1.3.1
SQLAlchemy==2.0.39
sqlparse==0.5.3
typing_extensions==4.12.2
yarl==1.18.3
